+++ generate/resolvers/index.js
@@ -11,6 +11,7 @@ function read(name) {
 
 const templates = {
   base: read('base'),
+  user: read('base'),
   fieldOfType: read('fieldOfType'),
   paginatedField: read('paginatedField'),
 };
@@ -22,7 +23,10 @@ function generateResolver(template) {
 }
 
 const generators = {
-  base({ typeName, TypeName }) {
+  base({ typeName, TypeName }, mode) {
+    if (mode === 'add-user') {
+      return templateToAst(templates.user, { typeName, TypeName });
+    }
     return templateToAst(templates.base, { typeName, TypeName });
   },
   belongsTo: generateResolver(templates.fieldOfType),
@@ -32,23 +36,23 @@ const generators = {
   hasAndBelongsToMany: generateResolver(templates.paginatedField),
 };
 
-export function generateResolversAst(inputSchema) {
+export function generateResolversAst(inputSchema, mode) {
   const type = inputSchema.definitions[0];
   const TypeName = type.name.value;
   const typeName = lcFirst(TypeName);
 
-  const ast = generators.base({ TypeName, typeName });
+  const ast = generators.base({ TypeName, typeName }, mode);
 
   // XXX: rather than hardcoding in array indices it would be less brittle to
   // walk the tree using https://github.com/benjamn/ast-types
-  const typeResolversAst = ast.program.body[0] // const
-    .declarations[0].init // object expression
-    .properties[0].value; // object value
+  const typeResolversAst =
+    ast.program.body[1].declarations[0].init.properties[0].value;
+    // const // object expression // object value
 
   generatePerField(type, generators).forEach((resolverFunctionAst) => {
-    const resolverProperty = resolverFunctionAst.program.body[0] // variable declaration
-      .declarations[0].init // object expression
-      .properties[0];
+    const resolverProperty =
+      resolverFunctionAst.program.body[0].declarations[0].init.properties[0];
+      // variable declaration // object expression
 
     typeResolversAst.properties.push(resolverProperty);
   });
@@ -56,7 +60,7 @@ export function generateResolversAst(inputSchema) {
   return ast;
 }
 
-export default function generateResolvers(inputSchema) {
-  const ast = generateResolversAst(inputSchema);
+export default function generateResolvers(inputSchema, mode) {
+  const ast = generateResolversAst(inputSchema, mode);
   return print(ast, { trailingComma: true }).code;
 }